+++ engine/src/main/java/nl/inl/blacklab/search/results/ResultsStats.java
@@ -76,7 +76,7 @@ public abstract class ResultsStats implements SearchResult {
      * @return
      */
     @Override
-    public int numberOfResultObjects() {
+    public long numberOfResultObjects() {
         return 1;
     }
